+++ src/http.c
@@ -356,7 +356,7 @@ int gg_http_watch_fd(struct gg_http *h)
 			unsigned int left;
 			char *line;
 
-			left = h->header_size - ((long)(tmp) - (long)(h->header) + sep_len);
+			left = h->header_size - ((size_t)(tmp) - (size_t)(h->header) + sep_len);
 
 			gg_debug(GG_DEBUG_MISC, "=> http, got all header "
 				"(%d bytes, %d left)\n",
